refactor: Change the path of 'LevelSet' related files to 'Utils'
