+++ Sources/Core/Emitter/VolumeGridEmitter3.cpp
@@ -11,8 +11,8 @@
 #include <Core/Emitter/VolumeGridEmitter3.hpp>
 #include <Core/Grid/CollocatedVectorGrid3.hpp>
 #include <Core/Grid/FaceCenteredGrid3.hpp>
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Surface/SurfaceToImplicit3.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 #include <Core/Utils/Macros.hpp>
 
 #include <utility>
