+++ Sources/Core/Solver/Grid/GridBackwardEulerDiffusionSolver3.cpp
@@ -8,9 +8,9 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Solver/FDM/FDMICCGSolver3.hpp>
 #include <Core/Solver/Grid/GridBackwardEulerDiffusionSolver3.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 
 namespace CubbyFlow
 {
