+++ Sources/Core/Solver/Grid/GridFluidSolver2.cpp
@@ -9,11 +9,11 @@
 // property of any third parties.
 
 #include <Core/Array/ArrayUtils.hpp>
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/SemiLagrangian/CubicSemiLagrangian2.hpp>
 #include <Core/Solver/Grid/GridBackwardEulerDiffusionSolver2.hpp>
 #include <Core/Solver/Grid/GridFluidSolver2.hpp>
 #include <Core/Solver/Grid/GridFractionalSinglePhasePressureSolver2.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 #include <Core/Utils/Logging.hpp>
 #include <Core/Utils/Timer.hpp>
