+++ Sources/Core/Solver/Grid/GridFractionalBoundaryConditionSolver2.cpp
@@ -9,10 +9,10 @@
 // property of any third parties.
 
 #include <Core/Array/ArrayUtils.hpp>
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Solver/Grid/GridFractionalBoundaryConditionSolver2.hpp>
 #include <Core/Surface/ImplicitSurface2.hpp>
 #include <Core/Surface/SurfaceToImplicit2.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 #include <Core/Utils/PhysicsHelpers.hpp>
 
 namespace CubbyFlow
