+++ Sources/Core/Solver/Grid/GridSinglePhasePressureSolver3.cpp
@@ -8,10 +8,10 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Solver/FDM/FDMICCGSolver3.hpp>
 #include <Core/Solver/Grid/GridBlockedBoundaryConditionSolver3.hpp>
 #include <Core/Solver/Grid/GridSinglePhasePressureSolver3.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 
 namespace CubbyFlow
 {
