+++ Sources/Core/Solver/LevelSet/LevelSetLiquidSolver2.cpp
@@ -9,10 +9,10 @@
 // property of any third parties.
 
 #include <Core/Grid/CellCenteredScalarGrid2.hpp>
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Solver/LevelSet/ENOLevelSetSolver2.hpp>
 #include <Core/Solver/LevelSet/FMMLevelSetSolver2.hpp>
 #include <Core/Solver/LevelSet/LevelSetLiquidSolver2.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 #include <Core/Utils/Logging.hpp>
 #include <Core/Utils/Timer.hpp>
