+++ Sources/Core/Solver/LevelSet/LevelSetLiquidSolver3.cpp
@@ -9,10 +9,10 @@
 // property of any third parties.
 
 #include <Core/Grid/CellCenteredScalarGrid3.hpp>
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Solver/LevelSet/ENOLevelSetSolver3.hpp>
 #include <Core/Solver/LevelSet/FMMLevelSetSolver3.hpp>
 #include <Core/Solver/LevelSet/LevelSetLiquidSolver3.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 #include <Core/Utils/Logging.hpp>
 #include <Core/Utils/Timer.hpp>
