+++ Sources/Core/Surface/CustomImplicitSurface3.cpp
@@ -8,8 +8,8 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Surface/CustomImplicitSurface3.hpp>
+#include <Core/Utils/LevelSetUtils.hpp>
 
 #include <utility>
